RecommendationsServer-Flask/requirements.txt
autopep8==1.6.0
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.0
gunicorn==20.1.0
importlib-metadata==4.11.3
iteration-utilities==0.11.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.4.2
pycodestyle==2.8.0
PyJWT==2.4.0
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
toml==0.10.2
Werkzeug==2.1.2
zipp==3.8.0
